:recycle: Complete moving transactions related files
