+++ framework/src/modules/http_api/controllers/node.js
@@ -18,6 +18,7 @@ const _ = require('lodash');
 const checkIpInList = require('../helpers/check_ip_in_list');
 const apiCodes = require('../api_codes');
 const swaggerHelper = require('../helpers/swagger');
+const { CACHE_KEYS_TRANSACTION_COUNT } = require('../../../components/cache');
 
 const { EPOCH_TIME, FEES } = global.constants;
 
@@ -42,6 +43,7 @@ function NodeController(scope) {
 	library = {
 		components: {
 			storage: scope.components.storage,
+			cache: scope.components.cache,
 		},
 		config: scope.config,
 		channel: scope.channel,
@@ -120,9 +122,12 @@ NodeController.getStatus = async (context, next) => {
 			secondsSinceEpoch,
 			loaded,
 			syncing,
-			transactions,
+			unconfirmedTransactions,
 			lastBlock,
 		} = await library.channel.invoke('chain:getNodeStatus');
+
+		// get confirmed count from cache or chain
+		const confirmed = await getConfirmedTransactionCount();
 
 		const networkHeight = await _getNetworkHeight();
 
@@ -135,7 +140,10 @@ NodeController.getStatus = async (context, next) => {
 			loaded,
 			networkHeight,
 			syncing,
-			transactions,
+			transactions: {
+				confirmed,
+				...unconfirmedTransactions,
+			},
 		};
 
 		return next(null, data);
@@ -324,6 +332,49 @@ async function _getNetworkHeight() {
 	);
 
 	return parseInt(networkHeight);
+}
+
+/**
+ * Get count of confirmedTransaction from cache
+ *
+ * @returns Number
+ * @private
+ */
+async function getConfirmedTransactionCount() {
+	// if cache is ready, then get cache and return
+	if (library.components.cache.cacheReady) {
+		try {
+			const { confirmed } = await library.components.cache.getJsonForKey(
+				CACHE_KEYS_TRANSACTION_COUNT
+			);
+			if (confirmed === undefined || confirmed === null) {
+				throw new Error(
+					'Transaction count wasn cached but confirmed did not exist'
+				);
+			}
+			return confirmed;
+		} catch (error) {
+			library.logger.warn("Transaction count wasn't cached", error);
+		}
+	}
+	const confirmed = await library.channel.invoke(
+		'chain:getConfirmedTransactionCount'
+	);
+	// only update cache if ready
+	if (library.components.cache.cacheReady) {
+		try {
+			await library.components.cache.setJsonForKey(
+				CACHE_KEYS_TRANSACTION_COUNT,
+				{
+					confirmed,
+				}
+			);
+		} catch (error) {
+			// Ignore error and just put warn
+			library.logger.warn("Transaction count wasn't cached", error);
+		}
+	}
+	return confirmed;
 }
 
 module.exports = NodeController;
